+++ src/components/modal/modal.module.css
@@ -1,5 +1,17 @@
+.modal {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
 .content {
   position: relative;
+  z-index: 1;
   min-width: 720px;
   min-height: 539px;
   background-color: #1c1c21;
@@ -8,6 +20,7 @@
     0px 16px 24px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04),
     0px 0px 1px rgba(0, 0, 0, 0.04);
   border-radius: 40px;
+  box-sizing: border-box;
 }
 
 .closeButton {
@@ -29,7 +42,4 @@
   bottom: 0;
   left: 0;
   background-color: rgba(0, 0, 0, 0.6);
-  display: flex;
-  justify-content: center;
-  align-items: center;
 }
